refactor: 집들이 리스트 필터 요소 UI 리팩토링 (작업중)
