+++ src/components/pages/ArticleList/filterContainer/FilterContainer.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect } from "react";
-import { useLocation } from "react-router-dom";
 import Filter from "./filter/Filter";
 import Dropdown from "../../../shared/dropdown/Dropdown";
 import { StFilterContainer, StFilteredInfoList } from "./styled";
@@ -51,42 +50,11 @@ const FilterContainer = () => {
         },
     }
 
-    const location = useLocation();
-
-    const handleQuery = (query, value) => {
-        console.log(query, value);
-    }
-
-    useEffect(() => {
-        console.log(location);
-    }, [location]);
-
     return (
         <>
         <StFilterContainer>
-            <Filter>정렬
-                <Dropdown>
-                    <ul>
-                        <li onClick={() => handleQuery('string', 'newest')}></li>
-                        <li onClick={() => handleQuery('string', 'newest')}>최근 인기순</li>
-                        <li onClick={() => handleQuery('string', 'newest')}>역대 인기순</li>
-                        <li onClick={() => handleQuery('string', 'newest')}>과거순</li>
-                    </ul>
-                </Dropdown>
-            </Filter>
-            <Filter>주거형태
-                <Dropdown>
-                    <ul>
-                        <li onClick={() => handleQuery('string', 'newest')}>원룸&오피스텔</li>
-                        <li onClick={() => handleQuery('string', 'newest')}>아파트</li>
-                        <li onClick={() => handleQuery('string', 'newest')}>빌라&연립</li>
-                        <li onClick={() => handleQuery('string', 'newest')}>단독주택</li>
-                        <li onClick={() => handleQuery('string', 'newest')}>사무공간</li>
-                        <li onClick={() => handleQuery('string', 'newest')}>상업공간</li>
-                        <li onClick={() => handleQuery('string', 'newest')}>기타</li>
-                    </ul>
-                </Dropdown>
-            </Filter>
+            <Filter filterInfo={filters.sorting} />
+            <Filter filterInfo={filters.residence} />
         </StFilterContainer>
         <StFilteredInfoList>
             <li><button>빌라 연립</button></li>
